Fourth commit
In this commit, I have done the bivariate analysis. Following type of analysis is done in this commit - Cancellation analysis, Revenue analysis, Trends in booking and ADR with respect to time (Year, month, week), and analysis of special request. Also, a pair plot has also been drawn in the end.